#51 Reorganized folder structure for screens to help keep it organized as the app grows.
