+++ chick-fil-a/src/config/routes/HomeStackNavigator.js
@@ -2,19 +2,20 @@ import { createStackNavigator } from 'react-navigation';
 import EStyleSheet from 'react-native-extended-stylesheet';
 
 import { colors, navigationHeaderTitle } from '../globalStyles';
+
+import { HomeScreen } from '../../screens';
 import {
-  HomeScreen,
+  GiveFeedbackContactForm,
+  GiveFeedbackForm,
+  GiveFeedbackList,
   HelpScreen,
   LegalScreen,
+  LicenseScreen,
   NotificationScreen,
   OptionsScreen,
-  QAScreen,
   PrivacyScreen,
-  GiveFeedbackList,
-  GiveFeedbackForm,
-  GiveFeedbackContactForm,
-  LicenseScreen
-} from '../../screens';
+  QAScreen
+} from '../../screens/about';
 
 const HomeStackNavigator = createStackNavigator(
   {
