+++ chick-fil-a/src/screens/about/index.js
@@ -1,0 +1,23 @@
+import GiveFeedbackContactForm from './GiveFeedbackContactForm';
+import GiveFeedbackForm from './GiveFeedbackForm';
+import GiveFeedbackList from './GiveFeedbackList';
+import HelpScreen from './HelpScreen';
+import LegalScreen from './LegalScreen';
+import LicenseScreen from './LicenseScreen';
+import NotificationScreen from './NotificationScreen';
+import OptionsScreen from './OptionsScreen';
+import PrivacyScreen from './PrivacyScreen';
+import QAScreen from './QAScreen';
+
+export {
+  GiveFeedbackContactForm,
+  GiveFeedbackForm,
+  GiveFeedbackList,
+  HelpScreen,
+  LegalScreen,
+  LicenseScreen,
+  NotificationScreen,
+  OptionsScreen,
+  PrivacyScreen,
+  QAScreen
+};
