+++ chick-fil-a/src/screens/index.js
@@ -1,35 +1,8 @@
-import HelpScreen from './HelpScreen';
+import FindScreen from './FindScreen';
 import HomeScreen from './HomeScreen';
 import IntroScreen from './IntroScreen';
-import FindScreen from './FindScreen';
-import OptionsScreen from './OptionsScreen';
 import MenuScreen from './MenuScreen';
 import MyCardScreen from './MyCardScreen';
 import MyOrderScreen from './MyOrderScreen';
-import QAScreen from './QAScreen';
-import LegalScreen from './LegalScreen';
-import NotificationScreen from './NotificationScreen';
-import PrivacyScreen from './PrivacyScreen';
-import GiveFeedbackList from './GiveFeedbackList';
-import GiveFeedbackForm from './GiveFeedbackForm';
-import GiveFeedbackContactForm from './GiveFeedbackContactForm';
-import LicenseScreen from './LicenseScreen';
 
-export {
-  HelpScreen,
-  HomeScreen,
-  IntroScreen,
-  FindScreen,
-  OptionsScreen,
-  MenuScreen,
-  MyCardScreen,
-  MyOrderScreen,
-  QAScreen,
-  LegalScreen,
-  NotificationScreen,
-  PrivacyScreen,
-  GiveFeedbackList,
-  GiveFeedbackForm,
-  GiveFeedbackContactForm,
-  LicenseScreen
-};
+export { HomeScreen, IntroScreen, FindScreen, MenuScreen, MyCardScreen, MyOrderScreen };
